+++ hai_core/src/traits/focusable.rs
@@ -1,5 +1,5 @@
 pub trait Focusable {
     /// check whether the input point is on the node-like,
     /// coordinate of the point is relative to its parent
-    fn contains(&self, x: i32, y: i32) -> bool;
+    fn contains(&self, x: f64, y: f64) -> bool;
 }
